tests moved to one folder
