--- employeeservice/src/test/resource/application-test.properties
@@ -1,9 +1,0 @@
-spring.h2.console.enabled=true
-
-spring.datasource.name=testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.url=jdbc:h2:file:./data/employeedb;SCHEMA=TestSchema
-spring.jpa.hibernate.ddl-auto=create
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
